move parseVatSlots.js out of src/vats/ up to src/
