+++ src/kernel/deviceManager.js
@@ -1,7 +1,7 @@
 import harden from '@agoric/harden';
 import { insist } from '../insist';
 import { insistKernelType } from './parseKernelSlots';
-import { insistVatType, parseVatSlot } from '../vats/parseVatSlots';
+import { insistVatType, parseVatSlot } from '../parseVatSlots';
 
 export default function makeDeviceManager(
   deviceName,
