+++ src/kernel/vatManager.js
@@ -2,7 +2,7 @@ import harden from '@agoric/harden';
 import djson from './djson';
 import { insist } from '../insist';
 import { insistKernelType, parseKernelSlot } from './parseKernelSlots';
-import { insistVatType, parseVatSlot } from '../vats/parseVatSlots';
+import { insistVatType, parseVatSlot } from '../parseVatSlots';
 
 export default function makeVatManager(
   vatID,
